+++ service/internal/cloudconfig/template/ephemeral_var_lib_containerd_mount.go
@@ -1,0 +1,12 @@
+package template
+
+const EphemeralVarLibContainerdMount = `
+[Unit]
+Description=Mount ephemeral volume on /var/lib/containerd
+[Mount]
+What=/dev/disk/by-label/containerd
+Where=/var/lib/containerd
+Type=xfs
+[Install]
+RequiredBy=local-fs.target
+`