+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Containerd EBS Volume.
+
+### Fixed
+
+- Fix `crictl.yaml` on worker nodes.
+
 ## [12.0.0] - 2022-07-14
 
 ### Added